+++ src/full_text/FullTextIndex.cpp
@@ -85,7 +85,7 @@ vector<FullTextResult> FullTextIndex::search_phrase(const string &phrase, int li
 
 	Profiler profiler2("value_intersection");
 	size_t shortest_vector;
-	vector<uint32_t> score_vector;
+	vector<float> score_vector;
 	vector<size_t> result_ids = value_intersection(result_map, shortest_vector, score_vector);
 
 	vector<FullTextResult> result;
@@ -112,7 +112,7 @@ vector<FullTextResult> FullTextIndex::search_phrase(const string &phrase, int li
 	Profiler profiler3("sorting results");
 	if (result.size() > limit) {
 		nth_element(score_vector.begin(), score_vector.begin() + (limit - 1), score_vector.end());
-		const uint32_t nth = score_vector[limit - 1];
+		const float nth = score_vector[limit - 1];
 
 		vector<FullTextResult> top_result;
 		for (const FullTextResult &res : result) {
@@ -177,7 +177,7 @@ void FullTextIndex::download(const SubSystem *sub_system) {
 }
 
 vector<size_t> FullTextIndex::value_intersection(const map<size_t, FullTextResultSet *> &values_map,
-	size_t &shortest_vector_position, vector<uint32_t> &scores) const {
+	size_t &shortest_vector_position, vector<float> &scores) const {
 	Profiler value_intersection("FullTextIndex::value_intersection");
 	
 	if (values_map.size() == 0) return {};
@@ -202,10 +202,10 @@ vector<size_t> FullTextIndex::value_intersection(const map<size_t, FullTextResul
 		bool all_equal = true;
 		uint64_t value = value_ptr[positions[shortest_vector_position]];
 
-		uint32_t score_sum = 0;
+		float score_sum = 0.0f;
 		for (auto &iter : values_map) {
 			const uint64_t *val_arr = iter.second->value_pointer();
-			const uint32_t *score_arr = iter.second->score_pointer();
+			const float *score_arr = iter.second->score_pointer();
 			const size_t len = iter.second->len();
 			size_t *pos = &(positions[iter.first]);
 			while (value > val_arr[*pos] && *pos < len) {
