+++ src/link_index/LinkIndexer.h
@@ -39,7 +39,7 @@ private:
 	vector<LinkShardBuilder *> m_shards;
 
 	void add_data_to_shards(uint64_t link_hash, const URL &source_url, const URL &target_url, const string &link_text,
-		uint32_t score);
+		float score);
 	void adjust_score_for_domain_link(const string &link_text, const Link &link);
 	void adjust_score_for_url_link(const string &link_text, const Link &link);
